+++ run_sim.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+    
+clear
+
+vvp output.out
+gtkwave dump.vcd